backend/requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
beanie==1.25.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==0.21.1
pytest==7.4.3
pydantic==2.7.2
